New slopes!
I’m trying out making the slopes from a 3D planar mesh in combination
with a tillable texture. They’re a bit jagged now but can be made much
smoother, should be a much more efficient way to generate these pieces
and make it so the sprites won’t have to be so high res.
Fixed a line in the PolygonTerrainGenerator that was preventing android
from building. Should build great now!