FX update
Added a ray caster to the Bear controller that detects the normal
direction of the terrain so that the Contact Emitter will tilt when its
on a hill. Transitions are a bit wonky right now, they’ll need to be
tuned.

Changed the camera in the art test scene to a perspective camera for
more accurate parallax especially when the bear is moving up\down.

Updated various prefabs. Terrain, Bear Body, etc

Added some set dressing to the slope mesh terrain pieces including BG
trees and FG snow slopes