First pass @ new camera
Created a new camera script that will follow the bear with an offset.
The offset is adjusted automatically when the bear is moving really
fast so that the camera zooms out and gives the bear more of a lead

Stubbed in a camera shake that returns the camera to its original
position after complete. I set it to trigger when the Avalanche is
near, but currently it triggers on collision with ANYTHING.

Will refine both tomorrow.