Lighting\Materials.
I’ve been messing with lighting and having dynamic time of day like in
alto. Right now the bear is backlit by the moon. Trying to get the
materials\shading to look correct no matter the lighting condition.

I had to lower the amount the camera zooms out when you are moving fast
because of the shadow bias. If the camera moves the shadows turn off
and its a bit jarring.