Bear HUD and Gameplay Loop
Implemented gameplay loop and HUD:

>Dashing or Smashing uses one of the boxes.
>When the slider refills, you get another smash/dash box
>Dashing/SMashing into trees gives you a bonus to your slider, so save
your dashes/smashes for big groups of trees.

Change the max number of boxes, the normal regen rate, and the bonus
regen for hitting trees on the Bear_Body Object. The code I wrote is
sort of disorganized, but ill clean it up if people like where this is
going.