week long update
about a weeks worth of changes here
>Added a temp cabin to try out some other collidable objects down the
mountain. it is made up of 3D objects and a rigid body kinematic
trigger that enables their movement on collision with the bear. The
bear also now has a spherical collider so that it will effect the
trajectory of the objects.
>Added more tree shapes to the tree gibs so that they are more
representative of the current tree prefab.
>Moved some of :pugs:’s mountain prefabs to the slopes that are being
generated so that we have endless BG generation. Its a bit clustered,
but with some adjustments I was able to make it so there is no popping
in and out as you slide down.
>Moved :pugs:’s BG prefab back a few units so that it is still visible
with the new BG generation, but it adds another level of parallax. If
we’re happy with the way the BG is being generated then we’ll have to
make BG changes within the slope prefabs in the future.
>Fixed a bug where the slopes were gradually generating off of the Z
axis and causing the bear to “float”.
>Slightly adjusted the lighting
>Reworked the TimedObjectDestructor script so that it won’t start
destroying until the object is active.